+++ axum-extra/src/routing/mod.rs
@@ -1,6 +1,6 @@
 //! Additional types for defining routes.
 
-use axum::{body::Body, handler::Handler, Router};
+use axum::{handler::Handler, Router};
 
 mod resource;
 
@@ -17,31 +17,6 @@ pub use self::typed::{FirstElementIs, TypedPath};
 
 /// Extension trait that adds additional methods to [`Router`].
 pub trait RouterExt<B>: sealed::Sealed {
-    /// Add the routes from `T`'s [`HasRoutes::routes`] to this router.
-    ///
-    /// # Example
-    ///
-    /// Using [`Resource`] which implements [`HasRoutes`]:
-    ///
-    /// ```rust
-    /// use axum::{Router, routing::get};
-    /// use axum_extra::routing::{RouterExt, Resource};
-    ///
-    /// let app = Router::new()
-    ///     .with(
-    ///         Resource::named("users")
-    ///             .index(|| async {})
-    ///             .create(|| async {})
-    ///     )
-    ///     .with(
-    ///         Resource::named("teams").index(|| async {})
-    ///     );
-    /// # let _: Router<axum::body::Body> = app;
-    /// ```
-    fn with<T>(self, routes: T) -> Self
-    where
-        T: HasRoutes<B>;
-
     /// Add a typed `GET` route to the router.
     ///
     /// The path will be inferred from the first argument to the handler function which must
@@ -151,13 +126,6 @@ impl<B> RouterExt<B> for Router<B>
 where
     B: axum::body::HttpBody + Send + 'static,
 {
-    fn with<T>(self, routes: T) -> Self
-    where
-        T: HasRoutes<B>,
-    {
-        self.merge(routes.routes())
-    }
-
     #[cfg(feature = "typed-routing")]
     fn typed_get<H, T, P>(self, handler: H) -> Self
     where
@@ -236,20 +204,6 @@ where
         P: TypedPath,
     {
         self.route(P::PATH, axum::routing::trace(handler))
-    }
-}
-
-/// Trait for things that can provide routes.
-///
-/// Used with [`RouterExt::with`].
-pub trait HasRoutes<B = Body> {
-    /// Get the routes.
-    fn routes(self) -> Router<B>;
-}
-
-impl<B> HasRoutes<B> for Router<B> {
-    fn routes(self) -> Router<B> {
-        self
     }
 }