+++ axum-macros/tests/typed_path/fail/route_not_starting_with_slash_non_empty.rs
@@ -1,0 +1,7 @@
+use axum_extra::routing::TypedPath;
+
+#[derive(TypedPath)]
+#[typed_path(":foo")]
+struct MyPath;
+
+fn main() {}