+++ axum/src/routing/url_params.rs
@@ -1,0 +1,45 @@
+use crate::util::{ByteStr, PercentDecodedByteStr};
+use http::Extensions;
+use matchit::Params;
+
+pub(crate) enum UrlParams {
+    Params(Vec<(ByteStr, PercentDecodedByteStr)>),
+    InvalidUtf8InPathParam { key: ByteStr },
+}
+
+pub(super) fn insert_url_params(extensions: &mut Extensions, params: Params) {
+    let current_params = extensions.get_mut();
+
+    if let Some(UrlParams::InvalidUtf8InPathParam { .. }) = current_params {
+        // nothing to do here since an error was stored earlier
+        return;
+    }
+
+    let params = params
+        .iter()
+        .filter(|(key, _)| !key.starts_with(super::NEST_TAIL_PARAM))
+        .map(|(key, value)| (key.to_owned(), value.to_owned()))
+        .map(|(k, v)| {
+            if let Some(decoded) = PercentDecodedByteStr::new(v) {
+                Ok((ByteStr::new(k), decoded))
+            } else {
+                Err(ByteStr::new(k))
+            }
+        })
+        .collect::<Result<Vec<_>, _>>();
+
+    match (current_params, params) {
+        (Some(UrlParams::InvalidUtf8InPathParam { .. }), _) => {
+            unreachable!("we check for this state earlier in this method")
+        }
+        (_, Err(invalid_key)) => {
+            extensions.insert(UrlParams::InvalidUtf8InPathParam { key: invalid_key });
+        }
+        (Some(UrlParams::Params(current)), Ok(params)) => {
+            current.extend(params);
+        }
+        (None, Ok(params)) => {
+            extensions.insert(UrlParams::Params(params));
+        }
+    }
+}