+++ axum/CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **added:** `Extension<_>` can now be used in tuples for building responses, and will set an
   extension on the response ([#797])
 - **added:** Implement `tower::Layer` for `Extension` ([#801])
+- **added:** `extract::Host` for extracting the hostname of a request ([#827])
+- **changed:** `Router::merge` now accepts `Into<Router>` ([#819])
 - **breaking:** `sse::Event` now accepts types implementing `AsRef<str>` instead of `Into<String>`
   as field values.
 - **breaking:** `sse::Event` now panics if a setter method is called twice instead of silently
@@ -59,6 +61,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **breaking:** `AddExtensionLayer` has been removed. Use `Extension` instead. It now implements
   `tower::Layer` ([#807])
 - **breaking:** `AddExtension` has been moved from the root module to `middleware`
+- **breaking:** `.nest("/foo/", Router::new().route("/bar", _))` now does the right thing and
+  results in a route at `/foo/bar` instead of `/foo//bar` ([#824])
+- **breaking:** Routes are now required to start with `/`. Previously routes such as `:foo` would
+  be accepted but most likely result in bugs ([#823])
 - **fixed:** Set `Allow` header when responding with `405 Method Not Allowed` ([#733])
 - **fixed:** Correctly set the `Content-Length` header for response to `HEAD`
   requests ([#734])
@@ -80,6 +86,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#800]: https://github.com/tokio-rs/axum/pull/800
 [#801]: https://github.com/tokio-rs/axum/pull/801
 [#807]: https://github.com/tokio-rs/axum/pull/807
+[#819]: https://github.com/tokio-rs/axum/pull/819
+[#823]: https://github.com/tokio-rs/axum/pull/823
+[#824]: https://github.com/tokio-rs/axum/pull/824
 
 # 0.4.4 (13. January, 2022)